gemified rails plugin
