+++ Gemfile
@@ -1,0 +1,5 @@
+source 'https://rubygems.org'
+
+gemspec
+
+gem 'activesupport', '~> 3.2.0'
