--- init.rb
@@ -1,2 +1,0 @@
-AbstractController::Base.send :extend,  BeforeRender       
-ActionController::Base.send :include,  BeforeRenderInstance
