+++ lib/rails3_before_render.rb
@@ -1,0 +1,8 @@
+require 'rails3_before_render/version'
+require 'active_support/concern'
+require 'rails3_before_render/before_render'
+require 'rails3_before_render/before_render_instance'
+
+module Rails3BeforeRender
+  
+end
