+++ lib/rails3_before_render/version.rb
@@ -1,0 +1,3 @@
+module Rails3BeforeRender
+  VERSION = "1.0.0"
+end
